Move debugger utils into debugger runtime module
