+++ bvm/ballerina-rt/build.gradle
@@ -117,6 +117,7 @@ dependencies {
     dist 'javax.transaction:javax.transaction-api:1.3'
 
     // debugger runtime helpers
+    dist project(':debug-adapter:debug-adapter-runtime')
     dist project(path: ':debug-adapter:debug-adapter-runtime', configuration: 'distributionBirJar')
 
     // Other Runtime Libs
