+++ misc/debug-adapter/modules/debug-adapter-runtime/build.gradle
@@ -18,9 +18,6 @@ apply from: "$rootDir/gradle/ballerinaNativeStdLibBuild.gradle"
 
 dependencies {
     implementation project(':ballerina-lang')
-    implementation project(':ballerina-lang:jballerina.java')
-    implementation project(':ballerina-lang:annotations')
-    implementation project(':ballerina-config')
     implementation project(':ballerina-runtime')
 }
