+++ misc/debug-adapter/modules/debug-adapter-runtime/src/main/java/module-info.java
@@ -1,0 +1,3 @@
+module org.ballerinalang.debugadapter.runtime {
+    requires io.ballerina.runtime;
+}
